api-server/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "uvicorn>=0.42.0",
    "pydantic>=2.0.0",
    "common_util",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.9",
]

[tool.uv.sources]
common_util = { path = "../common_util", editable = true }
